dts: adjust proximity near level
Adjust the proximity-near-level by one to account for an error in
calibration using iio-sensor-proxy, which interpreted the
in_proximity_nearlevel attribute as strictly-greater-than instead of
greater-than as it's specified in the kernel documentation.

It has been pointed out that devices with a DT calibrated against
iio-sensor-proxy will see a slight change in the interpretation of the near
level criterion.

Signed-off-by: Philipp Jungkamp <p.jungkamp@gmx.net>